+++ Source/GateRimSG1/Goauld/HediffCompProperties_GoauldSymbiote.cs
@@ -1,0 +1,12 @@
+using Verse;
+
+namespace GateRimSG1.Goauld
+{
+    public class HediffCompProperties_GoauldSymbiote : HediffCompProperties
+    {
+        public HediffCompProperties_GoauldSymbiote()
+        {
+            compClass = typeof(HediffComp_GoauldSymbiote);
+        }
+    }
+}